Bump spring-version from 2.5.5 to 2.6.1 in /officefloor/bom

+++ officefloor/tutorials/TransactionHttpServer/src/main/java/net/officefloor/tutorial/transactionhttpserver/Post.java
@@ -2,11 +2,10 @@ package net.officefloor.tutorial.transactionhttpserver;
 
 import java.io.Serializable;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -28,7 +27,7 @@ public class Post implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	private String content;
